pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["dbp"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
drf-yasg = "^1.21.7"
faker = "^30.0.0"
djangorestframework-simplejwt = "^5.3.1"
factory-boy = "^3.3.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest-django = "^4.9.0"
freezegun = "^1.5.1"
coverage = "^7.6.1"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
